Implement pwm locally
The simple_pwm in the avr-hal crate only really supports the main pwm
output, not the complementary one. To do so nicely would need some
bigger refactoring to ensure the main output can only be used as an
input if not used as a PWM output.

To solve that simpler implement a small pwm driver in this crate

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>